refactoring getTranslationFn
