+++ src/server/configs/shared/ql-chart-template.ts
@@ -4,14 +4,16 @@ import type {ServerI18n} from '../../../i18n/types';
 import type {QlExtendedConfig, StringParams} from '../../../shared';
 import {QLChartType, QL_TYPE, isMonitoringOrPrometheusChart} from '../../../shared';
 import {mapQlConfigToLatestVersion} from '../../../shared/modules/config/ql';
-import {getTranslationFn} from '../../utils/language';
+import {getTranslationFn} from '../../../shared/modules/language';
 
 export default {
     module: 'libs/qlchart/v1',
     identifyParams: (chart: QlExtendedConfig, req: Request) => {
         const i18nServer: ServerI18n = req.ctx.get('i18n');
 
-        const config = mapQlConfigToLatestVersion(chart, {i18n: getTranslationFn(i18nServer)});
+        const config = mapQlConfigToLatestVersion(chart, {
+            i18n: getTranslationFn(i18nServer.getI18nServer()),
+        });
         const {chartType, params} = config;
 
         const availableParams: StringParams = {};
@@ -44,7 +46,9 @@ export default {
     identifyChartType: (chart: QlExtendedConfig, req: Request) => {
         const i18nServer: ServerI18n = req.ctx.get('i18n');
 
-        const config = mapQlConfigToLatestVersion(chart, {i18n: getTranslationFn(i18nServer)});
+        const config = mapQlConfigToLatestVersion(chart, {
+            i18n: getTranslationFn(i18nServer.getI18nServer()),
+        });
 
         const {visualization, chartType} = config;
         const id = visualization.id;
@@ -78,7 +82,9 @@ export default {
     identifyLinks: (chart: QlExtendedConfig, req: Request) => {
         const i18nServer: ServerI18n = req.ctx.get('i18n');
 
-        const config = mapQlConfigToLatestVersion(chart, {i18n: getTranslationFn(i18nServer)});
+        const config = mapQlConfigToLatestVersion(chart, {
+            i18n: getTranslationFn(i18nServer.getI18nServer()),
+        });
         return {
             connection: config.connection.entryId,
         };
