+++ src/ui/units/ql/store/actions/ql.ts
@@ -4,6 +4,7 @@ import {i18n} from 'i18n';
 import _ from 'lodash';
 import type {match as Match} from 'react-router-dom';
 import {mapQlConfigToLatestVersion} from 'shared/modules/config/ql';
+import {getTranslationFn} from 'shared/modules/language';
 import type {
     QLConfigQuery,
     QlConfig,
@@ -660,6 +661,7 @@ export const initializeApplication = (args: InitializeApplicationArgs) => {
                 if (typeof loadedEntry.data?.shared === 'string') {
                     entry.data.shared = mapQlConfigToLatestVersion(
                         JSON.parse(loadedEntry.data.shared),
+                        {i18n: getTranslationFn(i18n)},
                     );
                 } else {
                     throw new Error(i18n('sql', 'error_failed-to-parse-loaded-chart'));
